Stop using non-blocking mode
When stdin and stdout refer to the same socket (eg. when called via
ssh), then setting non-blocking mode on one also causes non-blocking to
be set to the other. This makes writes to stdout non-blocking too, and
they can fail with EAGAIN.

Rather than having to cope with this, easier just to stop using
non-blocking mode as a workaround for Python not having a "block until
some data is available" mode and use os.read directly.